requirements.txt
accelerate
diffusers
invisible_watermark
torch
transformers
xformers
IPython
gradio
huggingface_hub
Pillow
PySimpleGUI
PySimpleGUIWeb 
python-dotenv